+++ operator/tools/controller-gen/go.mod
@@ -1,7 +1,8 @@
 module github.com/stackrox/rox/operator/tools/controller-gen
 
-go 1.23
-toolchain go1.23.2
+go 1.23.0
+
+toolchain go1.23.4
 
 require sigs.k8s.io/controller-tools v0.16.5
