+++ tests/performance/scale/go.mod
@@ -1,7 +1,8 @@
 module github.com/stackrox/stackrox/performance-scale-tests
 
-go 1.23
-toolchain go1.23.2
+go 1.23.0
+
+toolchain go1.23.4
 
 require (
 	github.com/cloud-bulldozer/go-commons v1.0.11
