+++ tools/test/go.mod
@@ -1,7 +1,8 @@
 module github.com/stackrox/stackrox/tools/test
 
-go 1.23
-toolchain go1.23.2
+go 1.23.0
+
+toolchain go1.23.4
 
 require (
 	github.com/jstemmer/go-junit-report/v2 v2.1.0
